Extract re-usable logic from UserForm into abstract View class and extend View to UserForm
