+++ src/views/UserForm.ts
@@ -1,16 +1,6 @@
-import { User } from '../models/User';
+import { View } from './View';
 
-export class UserForm {
-	constructor(public parent: Element, public model: User) {
-		this.bindModel();
-	}
-
-	bindModel(): void {
-		this.model.on('change', () => {
-			this.render();
-		});
-	}
-
+export class UserForm extends View {
 	eventsMap(): { [key: string]: () => void } {
 		return {
 			'click:.set-age': this.onSetRandomAge,
@@ -42,23 +32,5 @@ export class UserForm {
         <button class="set-age">Set Random Age</button>
       </div>
     `;
-	}
-
-	bindEvents(fragment: DocumentFragment): void {
-		const eventsMap = this.eventsMap();
-		for (let eventKey in eventsMap) {
-			const [eventName, selector] = eventKey.split(':');
-			fragment.querySelectorAll(selector).forEach(element => {
-				element.addEventListener(eventName, eventsMap[eventKey]);
-			});
-		}
-	}
-
-	render(): void {
-		this.parent.innerHTML = '';
-		const templateElement = document.createElement('template');
-		templateElement.innerHTML = this.template();
-		this.bindEvents(templateElement.content);
-		this.parent.append(templateElement.content);
 	}
 }
